[core] Resurrect removed methods/classes for maven-pmd-plugin compat
maven-pmd-plugin 3.21.0 is the latest version of the plugin, which
unfortunately does no longer support PMD 7.0.0-SNAPSHOT, since some API
that is used by the plugin is no longer available.

This prevents testing new PMD version in existing projects that make use
of the Maven plugin.

Resurrect the removed API to re-enable compatibility with the plugin.